Initial commit
Initial commit of current developent source code.

This code includes all functionality for a 'real' diversity module managed by  arduino .  So 'old' systems will require an upgrade on the hardware to benifit from this.

That being said.. code will work fine on the old systems; but diveristy will not be managed by the code in this repo.